Restructure m_viz notebook into application/notebooks/, split UQ code out
Reorganizes PR #490 uq-usecase contribution to match the existing
directory conventions (application/, tests/UnitTests/Area):

- application/notebooks/m_viz/: parsing.py, geo.py, plotting.py split
  out of the single 1220-line m_viz_utils.py by concern.
- application/notebooks/gridkit_common/: the two functions
  (attach_json_ids, get_case_path_for_editing) that gridkit_utils.py
  actually exercised from the notebook.
- application/notebooks/uq_sampling/: the UQ parameter-sampling,
  run-orchestration, and Genrou dispatch-patching code that was bundled
  into gridkit_utils.py despite being unrelated to and unused by the
  m_viz notebook. Kept in this PR per request, split into
  sampling.py, runs.py, dispatch.py, and documented as not yet wired to
  a notebook in this repo.
- application/notebooks/pyproject.toml: installable package, replacing
  the notebook sys.path.insert hack.
- tests/UnitTests/Python: pytest unit tests for m_viz.parsing and
  m_viz.geo (previously untested).
- Wired into the C++ build as opt-in: GridKit_ENABLE_NOTEBOOKS gates
  add_subdirectory(application/notebooks), which exposes a
  notebooks_pyenv custom target that creates a venv and pip-installs the
  package. GridKit_ENABLE_NOTEBOOKS also gates the ctest Python suite,
  which runs against that venv, mirroring the existing test_install
  ordering: build notebooks_pyenv, then ctest.
- Replaced hardcoded personal paths (under /Users/isatkaus and
  /home/isatkaus) in the notebook with environment-variable-driven
  config (GRIDKIT_CASE_DATA_DIR, GRIDKIT_REPO, GRIDKIT_FIGS_DIR),
  consistent with the existing NREL_CLUSTER/onkestrel pattern already in
  the notebook.
- Dropped _build_branch_line_arrays from m_viz_utils.py: confirmed dead
  code, never called from plot_grid or anywhere else.

Verified: all modules py_compile clean, notebook JSON is well-formed,
pip install -e application/notebooks succeeds in a clean venv, all 18
pytest tests pass, and the full CMake flow (configure with
-DGridKit_ENABLE_NOTEBOOKS=ON, build notebooks_pyenv, ctest) passes
end-to-end.
