+++ application/notebooks/gridkit_common/__init__.py
@@ -1,0 +1,5 @@
+"""Small helpers shared across application/notebooks notebooks."""
+
+from .ids import attach_json_ids, get_case_path_for_editing
+
+__all__ = ["attach_json_ids", "get_case_path_for_editing"]
